# docs/swagger.yaml
basePath: /api
definitions:
...
  request.CreateMailTemplateRequest:
    properties:
      body:
        type: string
      document_type_id:
        type: string
      name:
        type: string
      subject:
        type: string
    required:
    - body
    - document_type_id
    - name
    - subject
    type: object
...
  response.DocumentTypeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  response.MailTemplateResponse:
    properties:
      body:
        type: string
      created_at:
        type: string
      document_type:
        $ref: '#/definitions/response.DocumentTypeResponse'
      document_type_id:
        type: string
      id:
        type: string
      name:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Julong Recruitment API Docs
  version: "1.0"
paths:
...
  /mail-templates:
    get:
      consumes:
      - application/json
      description: Find all mail templates paginated
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: Created At
        in: query
        name: created_at
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Find all mail templates paginated
      tags:
      - Mail Templates
    post:
      consumes:
      - application/json
      description: Create mail template
      parameters:
      - description: Create Mail Template
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.CreateMailTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.MailTemplateResponse'
      security:
      - BearerAuth: []
      summary: Create mail template
      tags:
      - Mail Templates
  /mail-templates/{id}:
    delete:
      consumes:
      - application/json
      description: Delete mail template
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete mail template
      tags:
      - Mail Templates
    get:
      consumes:
      - application/json
      description: Find mail template by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Find mail template by ID
      tags:
      - Mail Templates
  /mail-templates/document-type/{id}:
    get:
      consumes:
      - application/json
      description: Find all mail templates by document type ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MailTemplateResponse'
      security:
      - BearerAuth: []
      summary: Find all mail templates by document type ID
      tags:
      - Mail Templates
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"